Move UuidCreateMethod
