+++ src/NullDev/Skeleton/SourceFactory/Uuid4IdentitySourceFactory.php
@@ -6,11 +6,11 @@ namespace NullDev\Skeleton\SourceFactory;
 
 use NullDev\Skeleton\Definition\PHP\Methods\ConstructorMethod;
 use NullDev\Skeleton\Definition\PHP\Methods\ToStringMethod;
-use NullDev\Skeleton\Definition\PHP\Methods\UuidCreateMethod;
 use NullDev\Skeleton\Definition\PHP\Parameter;
 use NullDev\Skeleton\Definition\PHP\Types\ClassType;
 use NullDev\Skeleton\Definition\PHP\Types\TypeDeclaration\StringType;
 use NullDev\Skeleton\Source\ClassSourceFactory;
+use NullDev\Skeleton\Uuid\Definition\PHP\Methods\UuidCreateMethod;
 use Ramsey\Uuid\Uuid;
 
 class Uuid4IdentitySourceFactory implements SourceFactory
