+++ tests/NullDev/Skeleton/CodeGenerator/PhpParser/Methods/UuidCreateGeneratorTest.php
@@ -5,13 +5,13 @@ declare(strict_types=1);
 namespace tests\NullDev\Skeleton\CodeGenerator\PhpParser\Methods;
 
 use NullDev\Skeleton\CodeGenerator\PhpParser\Methods\UuidCreateGenerator;
-use NullDev\Skeleton\Definition\PHP\Methods\UuidCreateMethod;
 use NullDev\Skeleton\Definition\PHP\Types\ClassType;
+use NullDev\Skeleton\Uuid\Definition\PHP\Methods\UuidCreateMethod;
 use PhpParser\BuilderFactory;
 
 /**
  * @covers \NullDev\Skeleton\CodeGenerator\PhpParser\Methods\UuidCreateGenerator
- * @covers \NullDev\Skeleton\Definition\PHP\Methods\UuidCreateMethod
+ * @covers \NullDev\Skeleton\Uuid\Definition\PHP\Methods\UuidCreateMethod
  * @group  integration
  */
 class UuidCreateGeneratorTest extends BaseOutputGeneratorTestCase
